apapun itu

+++ database/migrations/2025_12_02_020812_create_order_items_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
             $table->foreignId('product_id')->constrained('products')->onDelete('restrict');
             $table->integer('quantity');
-            $table->integer('price');
-            $table->integer('subtotal');
+            $table->decimal('price', 12, 2);
+            $table->decimal('subtotal', 12, 2);
             $table->timestamps();
 
             $table->index('order_id');
